requirements.txt
Django==5.1.2
pandas
numpy
requests
django-filter
djangorestframework
django-ulid
django-bootstrap5                                   
gunicorn
python-dotenv
whitenoise
django-admin-soft-dashboard==1.0.11
django-debug-toolbar==4.3.0
django-crispy-forms
django-widget-tweaks
django-select2
crispy-bootstrap4
psycopg2
django_heroku
dj-static
google-generativeai
google-cloud-vision
Pillow
pdf2image
chardet